[android] DRY abstraction for File Deletion menu actions
